+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 #Changelog
 
 
+### 0.0.9 (2016-05-29)
+
+- FocusInputOnLangTabSwitch submodule: ctrl+click on a language tab will activate all the language tabs of the same language
+- SuperUser check moved to PHP instead PHP+JavaScript
+
+
+
 ### 0.0.8 (2016-05-28)
 
 - new tweak: HoverSaveDropdown
@@ -14,10 +21,12 @@
 - load module CSS on ready() to make module CSS tweaks available right after install
 
 
+
 ### 0.0.6 (2016-05-27)
 
 - new tweaks: LoadCollapsedModuleInfos and fixScrollbarJump (Reno)
 - revised default tweaks
+
 
 
 ### 0.0.5 (2016-05-27)
@@ -26,10 +35,12 @@
 - various bugfixes
 
 
+
 ### 0.0.4 (2016-05-27)
 
 - added default admin theme tweaks
 - various bugfixes
+
 
 
 ### 0.0.3 (2016-05-27)
@@ -37,9 +48,11 @@
 - added Reno admin theme tweaks
 
 
+
 ### 0.0.2 (2016-05-26)
 
 - added AdminLongClickDuration feature
+
 
 
 ### 0.0.1 (2016-05-25)